move getVersion to helper
