+++ test/TestClaim.ts
@@ -1,6 +1,6 @@
 import { ethers, upgrades } from "hardhat";
 import { Signer, BigNumber, FixedNumber } from "ethers";
-import { takeSnapshot, revertToSnapShot, getRevertMsg, advanceTimeAndBlock } from "./helper";
+import { takeSnapshot, revertToSnapShot, getRevertMsg, advanceTimeAndBlock, getVerification } from "./helper";
 const { soliditySha3, hexToNumber, sha3 } = require("web3-utils");
 
 import { use } from "chai";
@@ -419,17 +419,6 @@ describe("HappyTokenPool", () => {
       }
     });
   });
-
-  function getVerification(password, account) {
-    let hash: string | Uint8Array = ethers.utils.keccak256(ethers.utils.toUtf8Bytes(password));
-    let hash_bytes = Uint8Array.from(Buffer.from(hash.slice(2), "hex"));
-    hash = hash_bytes.slice(0, 5);
-    hash = "0x" + Buffer.from(hash).toString("hex");
-    return {
-      verification: soliditySha3(hexToNumber(hash), account),
-      validation: sha3(account),
-    };
-  }
 
   async function getResultFromPoolFill(happyTokenPoolDeployed, creationParams) {
     await happyTokenPoolDeployed.fill_pool(...Object.values(creationParams));
