+++ test/helper.ts
@@ -1,5 +1,6 @@
 import { network, ethers } from "hardhat";
 import { providers } from "ethers";
+import { soliditySha3, hexToNumber, sha3 } from "web3-utils";
 
 /**
  * @param {number} time second
@@ -34,4 +35,15 @@ export async function advanceTimeAndBlock(time: number): Promise<providers.Block
   await advanceTime(time);
   await advanceBlock();
   return Promise.resolve(ethers.provider.getBlock("latest"));
+}
+
+export function getVerification(password, account) {
+  let hash: string | Uint8Array = ethers.utils.keccak256(ethers.utils.toUtf8Bytes(password));
+  let hash_bytes = Uint8Array.from(Buffer.from(hash.slice(2), "hex"));
+  hash = hash_bytes.slice(0, 5);
+  hash = "0x" + Buffer.from(hash).toString("hex");
+  return {
+    verification: soliditySha3(hexToNumber(hash), account),
+    validation: sha3(account),
+  };
 }
